Split out `isVideoPressItem`
